pyproject.toml
[project]
name = "stt"
version = "0.1.0"
description = "音声文字起こしWebアプリケーション - 複数のSTTモデルに対応"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # STTモデル関連
    "openai>=1.0.0",
    "google-cloud-speech>=2.20.0",
    "boto3>=1.28.0",
    "azure-cognitiveservices-speech>=1.30.0",
    "elevenlabs>=0.2.0",
    # Streamlit Web App
    "streamlit>=1.28.0",
    "sqlalchemy>=2.0.0",
    "pandas>=2.0.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    # Google Generative AI (Gemini)
    "google-genai>=0.2.4",
    # その他
    "python-dotenv>=1.0.0",
    "watchdog>=6.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
